Update logging message in KahlaPushService for clarity
Changed the log message from "Queuing" to "Pushing" to better reflect the actual action being taken when events are sent to users. This change aims to improve understanding and debugging of the push event process in the KahlaPushService. Additionally, a TODO comment was added to `ThreadsController` to indicate the need for unit tests.